requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.1
envs==1.3
django-storages==1.9.1
mysqlclient==1.4.6
